==> notebooks/02-preprocess_data.ipynb <==
from e_commerce_recommendation.configs.settings import BASE_DIR
import polars as pl

DATA_DIR = BASE_DIR / "data"
RAW_DATA_DIR = DATA_DIR / "raw_data"

usa_products = pl.read_csv(RAW_DATA_DIR / "usa/usa.csv")
uk_products = pl.read_csv(RAW_DATA_DIR / "uk/uk.csv")
canada_products = pl.read_csv(RAW_DATA_DIR / "canada/canada.csv")
india_products = pl.read_csv(RAW_DATA_DIR / "india/india.csv")

usa_products

usa_categories = pl.read_csv(RAW_DATA_DIR / "usa/usa_categories.csv", new_columns=["categoryId", "categoryName"])
usa_products.with_columns(pl.lit("usa").alias("country")).rename({"category_id": "categoryId"}).select(
    ["asin", "title", "imgUrl", "productURL", "stars", "reviews", "price", "isBestSeller", "boughtInLastMonth", "categoryId", "country"]
).write_csv(RAW_DATA_DIR / "usa/usa_products_with_category_id.csv")
usa_categories.write_csv(RAW_DATA_DIR / "usa/usa_categories.csv")

uk_categories = uk_products.select("categoryName").unique(maintain_order=True).with_columns(pl.lit(pl.Series(range(uk_products.select("categoryName").unique(maintain_order=True).height)), allow_object=True).alias('categoryId'))
uk_products.join(uk_categories, on="categoryName", how="left").with_columns(pl.lit("uk").alias("country")).drop("categoryName").select(
    ["asin", "title", "imgUrl", "productURL", "stars", "reviews", "price", "isBestSeller", "boughtInLastMonth", "categoryId", "country"]
).write_csv(RAW_DATA_DIR / "uk/uk_products_with_category_id.csv")
uk_categories.select("categoryId", "categoryName").write_csv(RAW_DATA_DIR / "uk/uk_categories.csv")

canada_categories = canada_products.select("categoryName").unique(maintain_order=True).with_columns(pl.lit(pl.Series(range(canada_products.select("categoryName").unique(maintain_order=True).height)), allow_object=True).alias('categoryId'))
canada_products.join(canada_categories, on="categoryName", how="left").with_columns(pl.lit("canada").alias("country")).drop("categoryName").select(
    ["asin", "title", "imgUrl", "productURL", "stars", "reviews", "price", "isBestSeller", "boughtInLastMonth", "categoryId", "country"]
).write_csv(RAW_DATA_DIR / "canada/canada_products_with_category_id.csv")
canada_categories.select("categoryId", "categoryName").write_csv(RAW_DATA_DIR / "canada/canada_categories.csv")

india_categories = india_products.select("categoryName").unique(maintain_order=True).with_columns(pl.lit(pl.Series(range(india_products.select("categoryName").unique(maintain_order=True).height)), allow_object=True).alias('categoryId'))
india_products.join(india_categories, on="categoryName", how="left").with_columns(pl.lit("india").alias("country")).drop("categoryName").select(
    ["asin", "title", "imgUrl", "productURL", "stars", "reviews", "price", "isBestSeller", "boughtInLastMonth", "categoryId", "country"]
).write_csv(RAW_DATA_DIR / "india/india_products_with_category_id.csv")
india_categories.select("categoryId", "categoryName").write_csv(RAW_DATA_DIR / "india/india_categories.csv")

usa_products.height + uk_products.height + canada_products.height + india_products.height
