==> notebooks/03_merge_tables.ipynb <==
from e_commerce_recommendation.configs.settings import BASE_DIR
import polars as pl

DATA_DIR = BASE_DIR / "data"
RAW_DATA_DIR = DATA_DIR / "raw_data"

usa_categories = pl.read_csv(RAW_DATA_DIR / "usa/usa_categories.csv")
uk_categories = pl.read_csv(RAW_DATA_DIR / "uk/uk_categories.csv")
canada_categories = pl.read_csv(RAW_DATA_DIR / "canada/canada_categories.csv")
india_categories = pl.read_csv(RAW_DATA_DIR / "india/india_categories.csv")
usa_products = pl.read_csv(RAW_DATA_DIR / "usa/usa_products_with_category_id.csv").join(
    usa_categories, on="categoryId", how="left"
).drop("categoryId")
uk_products = pl.read_csv(RAW_DATA_DIR / "uk/uk_products_with_category_id.csv").join(
    uk_categories, on="categoryId", how="left"
).drop("categoryId")
canada_products = pl.read_csv(RAW_DATA_DIR / "canada/canada_products_with_category_id.csv").join(
    canada_categories, on="categoryId", how="left"
).drop("categoryId")
india_products = pl.read_csv(RAW_DATA_DIR / "india/india_products_with_category_id.csv").join(
    india_categories, on="categoryId", how="left"
).drop("categoryId")

products = pl.concat([usa_products, uk_products, canada_products, india_products], how="vertical")
products

import polars as pl

all_categories = pl.concat([
    usa_categories.select("categoryName"),
    uk_categories.select("categoryName"),
    canada_categories.select("categoryName"),
    india_categories.select("categoryName"),
])

global_categories = (
    all_categories
    .with_columns(pl.col("categoryName").str.to_lowercase())
    .unique()
    .sort("categoryName")
    .with_row_count("global_category_id", offset=1)
)

global_categories

from vllm import LLM

llm = LLM(
    model='Qwen/Qwen3-VL-Embedding-2B',
    runner='pooling',
    max_model_len=4096,
    gpu_memory_utilization=0.85,
    quantization="mxfp4"
)

def generate_text_embeddings(texts, llm: LLM):
    prompts = [
        f"Represent the meaning of this product category: {t}"
        for t in texts
    ]
    outputs = llm.embed(prompts)
    return [o.outputs.embedding for o in outputs]

embeddings = generate_text_embeddings(global_categories['categoryName'].to_list(), llm)

import numpy as np
from sklearn.cluster import AgglomerativeClustering

X = np.array(embeddings)

clusterer = AgglomerativeClustering(
    n_clusters=None,
    distance_threshold=0.18,   # tune this
    metric="cosine",
    linkage="average"
)

labels = clusterer.fit_predict(X)


categories = global_categories.with_columns(
    pl.Series("cluster_id", labels)
).sort("cluster_id")
categories.write_csv(DATA_DIR / "processed_data/global_categories_with_clusters.csv")

canonical = (
    categories
    .group_by("cluster_id")
    .agg([
        pl.col("categoryName").first().alias("canonical_name"),
        pl.len().alias("variants")
    ])
    .with_row_count("global_category_id", offset=1)
)
canonical

from googletrans import Translator
import asyncio

to_translate = (
    canonical
    .select("canonical_name")
    .unique()
    .to_series()
    .to_list()
)

translator = Translator()

async def translate_all(texts):
    tasks = [translator.translate(t, src="hi", dest="en") for t in texts]
    results = await asyncio.gather(*tasks)
    return {t: r.text for t, r in zip(texts, results)}

translations = await translate_all(to_translate)


canonical = canonical.with_columns(
    pl.col("canonical_name")
      .map_elements(lambda x: translations.get(x, x))
      .alias("canonical_name_en")
)
canonical

data = products.with_columns(
    pl.col("categoryName").str.to_lowercase().alias("categoryName")
).join(
    categories.select(["categoryName", "cluster_id"]),
    on="categoryName", how="inner"
).join(
    canonical.select(["cluster_id", "canonical_name_en"]),
    on="cluster_id"
).drop("categoryName")
data

data.estimated_size("mb")

import glob
india_files = glob.glob(str(RAW_DATA_DIR / "india" / "**" / "*.jpg"), recursive=True)
india_files = [f.split("/")[-1].split(".")[0] for f in india_files]
usa_files = glob.glob(str(RAW_DATA_DIR / "usa" / "**" / "*.jpg"), recursive=True)
usa_files = [f.split("/")[-1].split(".")[0] for f in usa_files]
uk_files = glob.glob(str(RAW_DATA_DIR / "uk" / "**" / "*.jpg"), recursive=True)
uk_files = [f.split("/")[-1].split(".")[0] for f in uk_files]
canada_files = glob.glob(str(RAW_DATA_DIR / "canada" / "**" / "*.jpg"), recursive=True)
canada_files = [f.split("/")[-1].split(".")[0] for f in canada_files]

indian_df = pl.DataFrame({"asin": india_files}).with_columns(pl.lit("india").alias("country"))
usa_df = pl.DataFrame({"asin": usa_files}).with_columns(pl.lit("usa").alias("country"))
uk_df = pl.DataFrame({"asin": uk_files}).with_columns(pl.lit("uk").alias("country"))
canada_df = pl.DataFrame({"asin": canada_files}).with_columns(pl.lit("canada").alias("country"))
all_images_df = pl.concat([indian_df, usa_df, uk_df, canada_df], how="vertical")
all_images_df

final_df = data.join(all_images_df, on=["asin", "country"], how="inner")
final_df

final_df.write_parquet(DATA_DIR / "processed_data/products_with_canonical_categories.parquet")
